+++ spec/features/users_can_view_project_information_spec.rb
@@ -24,6 +24,11 @@ RSpec.feature "Users can view project information" do
     page_has_project_information_list_row(label: "Age range", information: "11 to 18")
     page_has_project_information_list_row(label: "School phase", information: "Secondary")
 
+    expect(page).to have_content("Trust details")
+    page_has_project_information_list_row(label: "Incoming trust name", information: "THE ROMERO CATHOLIC ACADEMY")
+    page_has_project_information_list_row(label: "UK Provider Reference Number", information: "10061021")
+    page_has_project_information_list_row(label: "Companies House number", information: "09702162")
+
     expect(page).to have_content("Local authority details")
     page_has_project_information_list_row(label: "Local authority", information: "West Placefield Council")
   end