+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   and age range.
 - The project information view displays the team lead.
 - Users must provide the UKPRN of the incoming trust when creating a project.
+- The project information view has a Trust details section, which displays the
+  incoming trust name, UKPRN, and Companies House number.
 
 ## [Release 1][release-1]